Update to Grails 5: remove configmap deprecation warnings & transition to maven central
