--- .travis.yml
@@ -1,19 +1,0 @@
-language: groovy
-sudo: false
-before_cache:
-- rm -f $HOME/.gradle/caches/modules-2/modules-2.lock
-cache:
-  directories:
-  - $HOME/.gradle/caches/
-  - $HOME/.gradle/wrapper/
-  - $HOME/.m2/repositories/
-before_script:
-- rm -rf target
-script: ./travis-build.sh
-env:
-  global:
-  - TERM=dumb
-  - secure: CroA/+s7hSoM6PtfkO9AhZt34GB74DypJIBEYKM8r/5tZeYT7+ERxa2Q8vkU+ur9RPqWR4D+08ZufPcBgf2Nw1gds5aqeUPgvH6vAHTfh99xPBuiWXlcrLtw0QyLfXxBURhzKZnHuWt8jhQBdwqQ198gt0o8xeEoeeSeXnMFYok40ExhW3o+Tso/961ZOSszxVEWV3cDuGy7zw3ba8XCEHD8qEBPkb1iCQYnYQGZCO8kcQ1vQLEUYcevWVRneJhtZWHfMyEFxytHpMcsPfuPwb94Mk7hRHoxsXRFFk1WDY9/YS0tULENrEEp08YrMoNWncqO8z1es/XrgYO6Z0nKbP/XMTNQ64HFC+sriSU3WndqZGmllGmFVahlwLi7Y/Z7ahzf6diTEvLjImNKrc1ssqSVW+qvOxyRLVblMB8ZC2Uvb0tJnXUjYj1LKbe2b923+Q+EZQyFv39u3Fs4l/7cJroINSW/drSqnbWk/y6AdEMlZJKffYzaWXpCjhavUBrokEVA5cOjRfLVoLLn87kLKT5rnooU+Y1bqqzuHBcergLHv6CYEjHroeV/7RiLNf8mbo34sL+sGEsRGODCYKRFN2w698oirJHVPtpNZ8V5vmiGCv2hKl4g4sOAjlCyTOpQJ6HIEPw1HqhjZqRMo0W7ijn19q5nmJApwiUYwU0aR7c=
-  - secure: jLAS3Y3+xjhYpXxoPH1uoFTtShE8GHNdQTiuMyL+qWibq4urdor51XZvJmkPJv3Sy7t2R3dQGovzmYHKCy6HwlOw82BCW5jMCRZLYbdO7sEWs+lj2+Sa/ThqvfmaSmA/e2wd/jZsC47laz0BQ3CRcspxQE6Q8C+i78xGdmTF224eyF5BWm9bhavvz1cxJ0akdxnG78xmtM1Ou8CrCZnDnRimIi9g2eVWUwK2DBXiQlkRTUUuggqOy7360WT7dWCaODgsAIk3/Q53IzoSqMK5LGKEH5ijiRfU00XxhTHoNPhlG+A9+wzfC64w2a48GHIKF/9UAqI2b4+naciB4K23u9HPgx6q1Vzt3HJFQjjC1fuL2qMSwI4l6D3GicmLaRxDlsUP3migYOPOfSLabcAIV82D9SJ8tMIAw7Nfjx6qPcholcheJEO5gvDmNXKOhGdqgk9k++WQYGmzTQSdJqR6wneQkVjkIOUGem8qNrrhD6IH2pDKUVZu8vHKRcaDk3EQe/S0+ThFR5jfYWMY9qGya/cTxjmF4ZfM1jbT5vpEhiPd5Tv48qLIwOaCVRxwlcOaSeGOi5yt7Efav+2GLlRN7ZeIm5O0jZKERDVTUE98kf5NF7zYMsBOdQUgOyghiqGpQ536RRoyv4L5sIJVE5FgtRcsVCMYky9Af4JenvrY3Y4=
-  - secure: e/ZUME8a5zgeCthlcDOXKlWS6dad/m60W/BVu0i/8MRHiOZZjIfCyA8CJJqDgD72JSP3u40Z6tEM5TsWeF3i7aYoVq4F0DoHEQ/WThKlIG/PdngtL8E34iZfOojpPZw3PoEAhwaxPi7EELQlWnIevGKL4b5Y8eWZGiUoTiU2g9P6PahjKsra8pAx/XYW/Yv32LNFx0IqcImOb8d4gofVpdNTzw/MjM/YSvBhWgHR0r5saSgE0t/H6CFLt/bUvjKBpKT/ucKb5JsGAjQ/1R9eOPiDtOJSr2QcXf8Ju16tAY3dKhmhOB6K/rYE6JkcS1s84IFUnfZiVBkek5aY04faalv4AEcbhLe1NRM6gWt1hNqVlC6gzzAAkWpaqHC7pYO7zCCDxiGtDntuoiIbeNjQzd9unjbUZaH/f5IYzGParwTD7DstrvMi0BSm9zi1lYEWXetS/le/Mf18LazDCLizVStXhpaLAcCkZZGRfM4cRgSkLOoYYwDkGa6IY1HQCBW2AalDsmIlIGOeucjZhdkT3TcnME4aBvKBSnuXC9RS1tBEZoKuSs3iaySJHlnj9+E9BGxl8E7wGNS/XaSk6qavejVPFAKVjAXTnE2MNH+dCXvwdxdhjI1UAB1TqJHb423ygJLRAblUQxWFSKf0W/yGBWLLm+SUScEE/ZBJ1eE9FP0=
-  - secure: ALHSOEOYgyHiniGJwxc02qB5jLFu7owQBlCrUTVobljY6qMzQK84V3OqacQZBehpflDGSXDlYI1flOIBI8q9BrFQP0TNbJqAEGEBft0kwPUaO/Pzi0Cqlomf4vaTEiGIkMDfcrxGE1lkqejEpZnAqyUUO7ewLMkHLXhwLs6lZIhSvGJJOyI1P2jgidsz6dPcqjuKiUQQRdA/4fVvJubFoSd0vWPmFM5DoHOHwguFeqByKZbBlx2ET4q4/gyBJnN9BHynUbj0T724Gb2U0nchrOePk2SF7YSivhE2A5m8KZNNP5Frisis7xERfOiDS95jD36LQl2bFBa9iHcMGaoSFlAKXwEfan2vmROpETjo+lhhY7LMcc5FTr8N4dURS+XP8xFxNLooI8gaqJlC9OKMbq5Ic5vft/z8jtjtkq3IMD7MUn2xKnpHF50kU/rodObsDExu8qXOD6qUtFvwglPflG6PUHIbRGDdjRKUM4tXNH0HjSjouuteVe2Fkusklbq71EOCX0VhbWT+6GPmNzM5f1PvrNvHr9H5HHYhEUtjZmUyGc7JkEGlsrpoUawvQUU4A9sn0c7W+ZdDgbgtvr/BDJDyR1wjBVA/CHqHFSXNZuEH1NMGTY4rC/StP5nHXYZH2kKaXdckND6BEEWHnoiVV5ux3lhldH5cCHyP+WV7xFo=
