+++ README.md
@@ -1,7 +1,7 @@
 # i18n-enums
 
-Successor for the Grails 2.x plugin. Now ported to Grails 3.x
+Successor for the Grails 3.x plugin. Now ported to Grails 5.x
 
 This plugin adds an annotation usable on Enums to easy add and implement the MessageSourceResolvable interface in an standard way throughout a project.
 
-Documentation can be found here: http://sbglasius.github.io/i18n-enums/
+Documentation can be found here: http://jdaugherty.github.io/i18n-enums/
