+++ docs/build.gradle
@@ -1,0 +1,44 @@
+buildscript {
+	repositories {
+		mavenLocal()
+		maven { url "https://repo.grails.org/grails/core" }
+	}
+	dependencies {
+		classpath 'org.ajoberstar:gradle-git-publish:2.1.3'
+	}
+}
+
+plugins {
+	id 'org.asciidoctor.jvm.convert' version '2.4.0'
+}
+
+apply plugin: 'org.ajoberstar.git-publish'
+
+gitPublish {
+	repoUri = 'git@github.com:jdaugherty/i18n-enums.git'
+	branch  = 'gh-pages'
+
+	// what to publish, this is a standard CopySpec
+	contents {
+		from "${buildDir}/docs"
+	}
+}
+
+asciidoctor {
+	baseDir file('src/asciidoc')
+	sourceDir file('src/asciidoc')
+	outputDir file("$buildDir/docs")
+	options doctype: 'book'
+	attributes 'source-highlighter': 'coderay',
+			toc                 : 'left',
+			toclevels           : '2',
+			'toc-title'         : 'Table of Contents',
+			icons               : 'font',
+			id                  : project.name + ':' + project.version,
+			idprefix            : '',
+			idseparator         : '-',
+			version             : version
+}
+
+// Rebuild docs before pushing
+gitPublishPush.dependsOn asciidoctor
