+++ gradle.properties
@@ -1,9 +1,16 @@
-grailsVersion=3.1.12
-gradleWrapperVersion=2.13
+title=I18n Enums
+projectVersion=5.0.0-SNAPSHOT
+projectDesc=Adds an annotation usable on Enums to easy add and implement the MessageSourceResolvable interface in an standard way throughout a grails project.
+projectUrl=https://github.com/jdaugherty/i18n-enums
+githubSlug=jdaugherty/i18n-enums
+githubBranch=main
+developers=Søren Berg Glasius, Burt Beckwith, James Daugherty
 
-pluginVersion=2.0.3-SNAPSHOT
+gradlePublishPlugin=1.3.0
+grailsVersion=5.0.1
+grailsGradlePluginVersion=5.0.1
+groovyVersion=3.0.7
+grailsDocsVersion=4.0.0
 
-userOrg=sbglasius
-websiteUrl=https://github.com/sbglasius/I18nEnumsGrails3Plugin
-issueTrackerUrl=https://github.com/sbglasius/I18nEnumsGrails3Plugin/issues
-vcsUrl=https://github.com/sbglasius/I18nEnumsGrails3Plugin
+org.gradle.daemon=true
+org.gradle.parallel=true
