--- gradle/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Thu Mar 05 09:04:40 PST 2015
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.1-bin.zip
