+++ plugin/build.gradle
@@ -1,0 +1,52 @@
+buildscript {
+	repositories {
+		mavenLocal()
+		maven { url "https://repo.grails.org/grails/core" }
+	}
+	dependencies {
+		classpath "org.grails:grails-gradle-plugin:$grailsGradlePluginVersion"
+	}
+}
+
+// Plugin version is set in the root/build.gradle
+group "org.grails.plugins"
+
+apply plugin: "eclipse"
+apply plugin: "idea"
+apply plugin: "org.grails.grails-plugin"
+apply plugin: "signing"
+apply plugin: "maven-publish"
+
+dependencies {
+	developmentOnly("org.springframework.boot:spring-boot-devtools")
+	compileOnly "io.micronaut:micronaut-inject-groovy"
+	console "org.grails:grails-console"
+	implementation "org.springframework.boot:spring-boot-starter-logging"
+	implementation "org.springframework.boot:spring-boot-starter-validation"
+	implementation "org.springframework.boot:spring-boot-autoconfigure"
+	implementation "org.grails:grails-core"
+	profile "org.grails.profiles:plugin"
+	testImplementation "io.micronaut:micronaut-inject-groovy"
+	testImplementation "org.grails:grails-web-testing-support"
+}
+
+bootRun {
+	jvmArgs('-Dspring.output.ansi.enabled=always')
+	sourceResources sourceSets.main
+}
+
+// Enable if you wish to package this plugin as a standalone application
+bootJar.enabled = false
+
+//this makes the groovy docs much more readable by shorting packages and providing links
+groovydoc {
+	excludes = ['**/*GrailsPlugin.groovy','**/Application.groovy' ]
+	link('http://download.oracle.com/javase/8/docs/api/', 'java.', 'org.xml', 'javax.', 'org.xml.')
+	link("https://docs.spring.io/spring/docs/4.2.x/javadoc-api/", 'org.springframework')
+	link('http://groovy.codehaus.org/api/',  'groovy.', 'org.codehaus.groovy.')
+	link('https://docs.grails.org/3.3.x/api',  'grails.', 'org.grails.')
+	link('https://testing.grails.org/latest/api',  'grails.testing.', 'org.grails.testing.')
+	link('http://gorm.grails.org/latest/hibernate/api/',  'grails.gorm', 'grails.orm', 'org.grails.datastore', 'org.grails.orm')
+}
+
+apply from: rootProject.file("gradle/publishing.gradle")
