+++ plugin/grails-app/conf/application.yml
@@ -1,0 +1,19 @@
+---
+grails:
+  profile: web-plugin
+  codegen:
+    defaultPackage: grails.plugin.i18nEnums
+info:
+    app:
+        name: '@info.app.name@'
+        version: '@info.app.version@'
+        grailsVersion: '@info.app.grailsVersion@'
+
+---
+grails:
+  doc:
+    title: "I18nEnum Plugin"
+    subtitle: "Keep your Enums DRY"
+    authors: "Søren Berg Glasius, James Daugherty"
+    license: "Apache2.0 license"
+    copyright: "Copies of this document may be made for your own use and for distribution to others, provided that you do not charge any fee for such copies and further provided that each copy contains this Copyright Notice, whether distributed in print or electronically."
