+++ plugin/src/main/groovy/grails/plugins/i18nEnums/DefaultNameCase.groovy
@@ -1,0 +1,12 @@
+package grails.plugins.i18nEnums
+
+import groovy.transform.CompileStatic
+
+@CompileStatic
+enum DefaultNameCase {
+	UPPER_CASE,
+	LOWER_CASE,
+	CAPITALIZE,
+	ALL_CAPS,
+	UNCHANGED
+}
