+++ plugin/src/test/groovy/grails/plugins/i18nEnums/annotations/DefaultAnnotatedEnumSpec.groovy
