--- src/main/groovy/grails/plugins/i18nEnums/DefaultNameCase.groovy
@@ -1,9 +1,0 @@
-package grails.plugins.i18nEnums
-
-enum DefaultNameCase {
-	UPPER_CASE,
-	LOWER_CASE,
-	CAPITALIZE,
-	ALL_CAPS,
-	UNCHANGED
-}
