--- src/main/groovy/grails/plugins/i18nEnums/annotations/I18nEnum.groovy
@@ -1,22 +1,0 @@
-package grails.plugins.i18nEnums.annotations
-
-import org.codehaus.groovy.transform.GroovyASTTransformationClass
-import grails.plugins.i18nEnums.DefaultNameCase
-
-import java.lang.annotation.ElementType
-import java.lang.annotation.Retention
-import java.lang.annotation.RetentionPolicy
-import java.lang.annotation.Target
-
-@Retention(RetentionPolicy.SOURCE)
-@Target(ElementType.TYPE)
-@GroovyASTTransformationClass("grails.plugins.i18nEnums.transformation.I18nEnumTransformation")
-public @interface I18nEnum {
-    String prefix() default ''
-
-    String postfix() default ''
-
-    boolean shortName() default false
-
-    DefaultNameCase defaultNameCase() default DefaultNameCase.UNCHANGED
-}
